Se arregla la incompatibilidad de datos de ReseniaRepository que afectaba a ReseniaServiceImpl

+++ mscv-resenia/src/main/java/com/smedinamsvc/resenia/repository/ReseniaRepository.java
@@ -8,9 +8,10 @@ import org.springframework.stereotype.Repository;
 import com.smedinamsvc.resenia.model.Resenia;
 
 @Repository
-public interface ReseniaRepository extends JpaRepository<Resenia, Integer> {
+public interface ReseniaRepository extends JpaRepository<Resenia, Long> {
 
     // Método para buscar todas las reseñas de un producto específico
     List<Resenia> findByProductoId(Long productoId);
 }
 
+///Siempre revisar el JPA para que sean compatibles los datos. Eso era lo que causaba problema en el Service impl
